+++ tests/dummy/app/controllers/components-examples/flexberry-groupedit/settings-example.js
@@ -114,6 +114,15 @@ export default EditFormController.extend({
   deleteButton: true,
 
   /**
+    Flag: indicates whether 'flexberry-objectlistview' component is in 'defaultSortingButton' mode or not
+
+    @property defaultSortingButton
+    @type Boolean
+    @default true
+  */
+  defaultSortingButton: true,
+
+  /**
     Flag for 'flexberry-groupedit' component 'allowColumnResize' property.
 
     @property allowColumnResize
@@ -201,6 +210,7 @@ export default EditFormController.extend({
     '  tableStriped=tableStriped<br>' +
     '  createNewButton=createNewButton<br>' +
     '  deleteButton=deleteButton<br>' +
+    '  defaultSortingButton=defaultSortingButton<br>' +
     '  allowColumnResize=allowColumnResize<br>' +
     '  showAsteriskInRow=showAsteriskInRow<br>' +
     '  showCheckBoxInRow=showCheckBoxInRow<br>' +
@@ -297,6 +307,12 @@ export default EditFormController.extend({
       settingType: 'boolean',
       settingDefaultValue: false,
       bindedControllerPropertieName: 'showDeleteButtonInRow'
+    });
+    componentSettingsMetadata.pushObject({
+      settingName: 'defaultSortingButton',
+      settingType: 'boolean',
+      settingDefaultValue: true,
+      bindedControllerPropertieName: 'defaultSortingButton'
     });
     componentSettingsMetadata.pushObject({
       settingName: 'showEditMenuItemInRow',