+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * `flexberry-lookup` component:
     * Dropdown mode uses projection from `projection` property.
 
+### Added
+* List components:
+    * Add option to hide default sorting button.
+
 ## [2.1.0-beta.3] - 2019-03-15
 ### Added
 * `flexberry-lookup` component: